first commit, beginning progress
reasonable functionality so far for most of the buttons and the display.
Styling also in a decent spot, meant to look like a windows 95 app.

in the process of finishing both the
pos/neg switch and the parenthesis buttons.

next step is to refine the display to be more similar in working to
the app on my phone.

following that, the next step will be adding in correct functionality
fir the "%" button.